Move numba_njit function to jit.py
