+++ phylokit/balance.py
@@ -3,11 +3,11 @@ import math
 
 import numpy as np
 
-from . import core
+from . import jit
 from . import util
 
 
-@core.numba_njit
+@jit.numba_njit
 def _sackin_index(virtual_root, left_child, right_sib):
     stack = []
     root = left_child[virtual_root]
@@ -43,7 +43,7 @@ def sackin_index(ds):
     return _sackin_index(-1, ds.node_left_child.data, ds.node_right_sib.data)
 
 
-@core.numba_njit
+@jit.numba_njit
 def _colless_index(postorder, left_child, right_sib):
     num_leaves = np.zeros_like(left_child)
     total = 0.0
@@ -87,7 +87,7 @@ def colless_index(ds):
     )
 
 
-@core.numba_njit
+@jit.numba_njit
 def _b1_index(postorder, left_child, right_sib, parent):
     max_path_length = np.zeros_like(postorder)
     total = 0.0
@@ -121,7 +121,7 @@ def b1_index(ds):
     )
 
 
-@core.numba_njit
+@jit.numba_njit
 def general_log(x, base):
     """
     Compute the logarithm of x in base `base`.
@@ -134,7 +134,7 @@ def general_log(x, base):
     return math.log(x) / math.log(base)
 
 
-@core.numba_njit
+@jit.numba_njit
 def _b2_index(virtual_root, left_child, right_sib, base):
     root = left_child[virtual_root]
     stack = [(root, 1)]
